xtensor_forward.hpp cleanup
